% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_polyfun_est.R
\name{setup_polyfun_est}
\alias{setup_polyfun_est}
\title{Check that susieR is available for polyfun_est}
\usage{
setup_polyfun_est()
}
\value{
Invisible TRUE if susieR is available.
}
\description{
Check that susieR is available for polyfun_est
}
