% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_sbayesrc.R
\name{setup_sbayesrc}
\alias{setup_sbayesrc}
\title{Check that the dependencies of sbayesrc are available}
\usage{
setup_sbayesrc()
}
\value{
Invisible TRUE.
}
\description{
The implementation uses base R only.
}
