% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_susie.R
\name{setup_susie}
\alias{setup_susie}
\title{Check and install susieR}
\usage{
setup_susie()
}
\value{
Invisible TRUE if susieR is available.
}
\description{
Verifies that the susieR package is available. If not, attempts to install
it from CRAN.
}
