% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_polyfun_ldsc.R
\name{setup_polyfun_ldsc}
\alias{setup_polyfun_ldsc}
\title{Check that susieR is available for polyfun_ldsc}
\usage{
setup_polyfun_ldsc()
}
\value{
Invisible TRUE if susieR is available.
}
\description{
Check that susieR is available for polyfun_ldsc
}
