% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_sbayesrc.R
\name{setup_sbayesrc}
\alias{setup_sbayesrc}
\title{Check that dependencies for sbayesrc are available}
\usage{
setup_sbayesrc()
}
\value{
Invisible TRUE.
}
\description{
Check that dependencies for sbayesrc are available
}
